>> genomic_tools VERSION 2.4.0:
[1] ** BUG ** 
genomic_scans count, option -r:
A small number of windows that overlap with reference regions was skipped, because the query region index was not advanced properly in GenomicRegionSetScanner::Next(GenomicRegionSet *Ref).   


[2] ** NEW OPTIONS ** 
genomic_scans counts:
This operation now allows unsorted inputs. Options -S and -Sref can now be used if input regions and/or reference regions (option -r) are sorted.
  

[3] ** BUG ** 
GenomicRegionSetScannerSorted:
When ignore_strand is set, the algorithm was expecting an input with strand=='+' only. As a result, when instances with '-' were present, the output was incorrect. 
Now this is fixed, and strand information is properly ignored.  


[4] ** NEW CLASS ** 
UnsortedGenomicRegionSetScanner